Adding directives to selectively disable comment parsing.
This change adds so-called 'directives' to our comments.  A directive is
a comment line whose only contents are a hash, followed by the directive
name, e.g.

```
;; #DirectiveName
```

Directives are not included in the `*comments*` vector in any case.

The two directives implemented here are "#MargDisable" and
"#MargEnable".  The former, when encountered, causes comments read by
`read-comment` to be ignored.  The latter re-enables it.

The rationale for this change is to accommodate e.g. license-header
boilerplate without polluting the generated docs.