Add a new option to exclude source files from being processed by Marginalia.
The new option is available at the command line with "-e" or "--exclude". It can also be part of the project.clj file via the ":exclude" keyword.

I was requiring this feature because I often have local testing files which I don't want to be included in the documentation.

The new option is quite simple: it simply filter out the files that have been identified by the option out of the "sources" var before it is being used by multidoc! or uberdoc!