Add option to exclude lifted comments from source
This adds tests for the -L --lift-inline-comments option as a baseline, then it adds a new -X --exclude-lifted-comments option and tests for that, as well as implementing it in the parser (we exclude lifted comments when merging an adjacent sequence of merged comments and a block of code).

This also corrects a bug in inline comment processing where ; comments were also lifted (they should have been ignored / left alone).