Naive solution to #81
This works for docstrings that are flush left or indented two spaces, which should be the two common cases. The naivety is in assuming that a newline followed by a non-whitespace character indicates flush left, and also that it is safe to just strip the first two spaces off each line of the docstring. It's as safe as some of the other text manipulation Marginalia does :)

Caveat: If a docstring is indented three spaces (per example 2 in my comment on #81) then they'll be unaffected by this change (and they'll still get the weird formatting they got before!).